+++ api/src/api/newsletter/newsletter.controller.ts
@@ -1,22 +1,24 @@
-import { Controller, Post, Get, Put, Delete, HttpCode, HttpStatus, Body, Param } from '@nestjs/common';
-import { ApiBearerAuth, ApiUseTags, ApiOperation, ApiResponse, ApiImplicitQuery, ApiImplicitHeader } from '@nestjs/swagger';
+import { Controller, Post, Body} from '@nestjs/common';
+import { ApiBearerAuth, ApiUseTags, ApiOperation, ApiResponse } from '@nestjs/swagger';
 import { NewsletterModel } from './newsletter.model';
-import { NewsletterDTO } from './newsletter.dto';
 import { NewsletterService } from './newsletter.service';
+import { InjectModel } from 'nestjs-typegoose';
+import { ModelType } from 'typegoose';
 
 @Controller('newsletter')
 @ApiUseTags('Newsletters')
 export class NewsletterController {
 
-    constructor(private readonly newsletterService: NewsletterService) { }
+    constructor(private readonly newsletterService: NewsletterService,
+                @InjectModel(NewsletterModel) private readonly newsletterModel: ModelType<NewsletterModel>,
+      ) { }
 
     @Post()
     @ApiBearerAuth()
     @ApiOperation({ title: 'Create new newsletter' })
     @ApiResponse({ status: 201, description: 'Return newsletter.' })
-    @ApiResponse({ status: 404, description: 'Not Found.' })
-    @ApiResponse({ status: 403, description: 'Forbidden.' })
-    async create(@Body() newsletter: NewsletterDTO): Promise<NewsletterModel> {
-        return this.newsletterService.insert(newsletter);
+    async newsletter(@Body() body) {
+        const contact = new this.newsletterModel({ email: body.email });
+        await contact.save();
     }
 }
