+++ api/src/api/newsletter/newsletter.service.ts
@@ -9,14 +9,6 @@ export class NewsletterService {
 
     constructor(@InjectModel(NewsletterModel) private readonly newsletterModel: ModelType<NewsletterModel>) { }
 
-    async findAll(): Promise<NewsletterModel[]> {
-      return this.newsletterModel.find({}).exec();
-    }
-
-    async findById(id: string): Promise<NewsletterModel> {
-      return this.newsletterModel.findById(id).exec();
-    }
-
     async insert(newsletter: NewsletterDTO): Promise<NewsletterModel> {
       const createdNewsletter = new this.newsletterModel(newsletter);
       return createdNewsletter.save();
