+++ ui/src/app/shared/models/newsletters.model.ts
@@ -1,6 +1,6 @@
 export class Newsletters {
     constructor(
         // tslint:disable-next-line: variable-name
-        public _id: string,
+        public email: string,
     ) { }
 }
